Python waits for Python Debug Adapter
